+++ src/shaders/FX/kernel.frag
@@ -15,7 +15,7 @@ W_UBO_DEF{
 
 W_PC_DEF{
   UboObject ubo;
-  uint8_t s_InputColTex;
+  uint16_t s_InputColTex;
 }
 
 //
